WaveVisual update

+++ vrx_ign/src/WaveVisual.cc
@@ -255,8 +255,8 @@ void WaveVisualPrivate::OnUpdate()
         // RenderUti stores gazebo-entity user data as int
         // \todo(anyone) Change this to uint64_t in Gazebo H?
         auto variant = n->UserData("gazebo-entity");
-        const int *value = std::get_if<int>(&variant);
-        if (value && *value == static_cast<int>(this->entity))
+        const uint64_t *value = std::get_if<uint64_t>(&variant);
+        if (value && *value == static_cast<uint64_t>(this->entity))
         {
           this->visual = std::dynamic_pointer_cast<rendering::Visual>(n);
           break;
